Extracted functionality to library crate.
